add createGroupQueryReactor abstraction
